vote componentizado
